Adding a cast for VS 2008.  Fixing compiler warnings about unused variables.
git-svn-id: https://wush.net/svn/mdaus-oss/trunk@535 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/scene/source/AdjustableParams.cpp
@@ -30,7 +30,7 @@ namespace scene
 {
 AdjustableParams::AdjustableParams()
 {
-    std::fill_n(mParams, NUM_PARAMS, 0.0);
+    std::fill_n(mParams, static_cast<size_t>(NUM_PARAMS), 0.0);
 }
 
 std::string AdjustableParams::name(ParamsEnum param)
